fix(#1618): complete queue spec file split and fix review issues
- Remove unused QueryType variable from queueManipulation.dedup.spec.ts
- Fix weak test assertion in autoplay.spec.ts (toBeGreaterThanOrEqual → toBeGreaterThan)
- Unwrap conditional test assertions in priority.spec.ts to ensure they always run
- Extract scoring-related tests into new queueManipulation.scoring.spec.ts:
  - getGenreFamilies tests
  - calculateGenreFamilyPenalty tests
  - enrichWithAudioFeatures tests

This reduces priority.spec.ts from 816 to 683 lines, moving closer to the 500-line target.
